refactor: :truck: 네이밍 변경 FilterBottomSheet -> BeerFilterBottomSheet
