--- src/components/filter/FilterBottomSheet/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterBottomSheet';
